refactor: move tests to cassandra namespace
